# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  username: string;
  displayName: string;
  password: string;
  /** @nullable */
  avatar?: string | null;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatar?: string | null;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateStatusBody {
  isOnline: boolean;
}

export type ConversationType =
  (typeof ConversationType)[keyof typeof ConversationType];

export const ConversationType = {
  direct: "direct",
  group: "group",
} as const;

export interface Conversation {
  id: number;
  type: ConversationType;
  /** @nullable */
  name?: string | null;
  participants: User[];
  createdAt: string;
}

export type ConversationSummaryType =
  (typeof ConversationSummaryType)[keyof typeof ConversationSummaryType];

export const ConversationSummaryType = {
  direct: "direct",
  group: "group",
} as const;

export interface LinkPreview {
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface Reaction {
  id: number;
  messageId: number;
  userId: number;
  emoji: string;
  user?: User;
  createdAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  sender?: User;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  linkPreview?: LinkPreview;
  reactions: Reaction[];
  createdAt: string;
}

export interface ConversationSummary {
  id: number;
  type: ConversationSummaryType;
  /** @nullable */
  name?: string | null;
  otherUser?: User;
  lastMessage?: Message;
  unreadCount: number;
  updatedAt: string;
}

export interface CreateConversationBody {
  userId: number;
}

export interface SendMessageBody {
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface AddReactionBody {
  emoji: string;
}

export interface UploadResponse {
  url: string;
}

export type SearchUsersParams = {
  q: string;
};

export type ListMessagesParams = {
  limit?: number;
  before?: number;
};

export type UploadImageBody = {
  file: Blob;
};