# lib/api-zod/src/generated/types/conversationSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationSummaryType } from "./conversationSummaryType";
import type { Message } from "./message";
import type { User } from "./user";

export interface ConversationSummary {
  id: number;
  type: ConversationSummaryType;
  /** @nullable */
  name?: string | null;
  otherUser?: User;
  lastMessage?: Message;
  unreadCount: number;
  updatedAt: string;
}