# lib/api-zod/src/generated/types/loginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginBody {
  username: string;
  password: string;
}