# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { LinkPreview } from "./linkPreview";
import type { Reaction } from "./reaction";
import type { User } from "./user";

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  sender?: User;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  linkPreview?: LinkPreview;
  reactions: Reaction[];
  createdAt: string;
}