# lib/api-zod/src/generated/types/reaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Reaction {
  id: number;
  messageId: number;
  userId: number;
  emoji: string;
  user?: User;
  createdAt: string;
}