# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  username: string;
  displayName: string;
  password: string;
  /** @nullable */
  avatar?: string | null;
}