# lib/api-zod/src/generated/types/sendMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageBody {
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}