# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TeleChat messaging PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatar?: string | null;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}